+++ utils/file.js
@@ -6,6 +6,68 @@ const http = require('http');
 const https = require('https');
 
 /**
+ * Converts a string object to a JSON/YAML object.
+ * @param str String.
+ * @param options Parse options
+ * @returns {Promise<*>} Object data.
+ */
+async function parseString(str, options = {}) {
+  // Convert large number value safely before parsing
+  const encodedContent = encodeLargeNumbers(str);
+
+  // Default to YAML format
+  const toYaml = options.format !== 'json' && (!options.hasOwnProperty('json') || options.json !== true);
+
+  if (toYaml) {
+    try {
+      const obj = yaml.parse(encodedContent);
+      if (typeof obj === 'object') {
+        return obj
+      } else {
+        return SyntaxError('Invalid YAML');
+      }
+    } catch (yamlError) {
+      return yamlError;
+    }
+  } else {
+    try {
+      // Try parsing as JSON
+      return JSON.parse(encodedContent);
+    } catch (jsonError) {
+      return jsonError;
+    }
+  }
+}
+
+async function isJSON(str) {
+  try {
+    JSON.parse(str);
+    return true
+  } catch (e) {
+    return false;
+  }
+}
+
+async function isYaml(str) {
+  try {
+    const rest = yaml.parse(str);
+    return (typeof rest === 'object')
+  } catch (e) {
+    return false;
+  }
+}
+
+async function detectFormat(str) {
+  if (await isJSON(str) !== false) {
+    return 'json';
+  } else if (await isYaml(str) !== false) {
+    return 'yaml';
+  } else {
+    return 'unknown';
+  }
+}
+
+/**
  * Parse a JSON/YAML file and returns the parsed object
  * @param filePath Path to the JSON/YAML file.
  * @returns {Promise<unknown>} Parsed data object.
@@ -13,7 +75,7 @@ const https = require('https');
 async function parseFile(filePath) {
   try {
     const isRemoteFile = filePath.startsWith('http://') || filePath.startsWith('https://');
-    const isYamlFile = filePath.endsWith('.yaml') || filePath.endsWith('.yml');
+    // const isYamlFile = filePath.endsWith('.yaml') || filePath.endsWith('.yml');
 
     let fileContent;
     if (isRemoteFile) {
@@ -22,13 +84,9 @@ async function parseFile(filePath) {
       fileContent = await getLocalFile(filePath);
     }
 
-    // Convert large number value safely before parsing
-    const encodedContent = encodeLargeNumbers(fileContent);
+    // Encode & Parse file content
+    return parseString(fileContent)
 
-    // Parse file content
-    return isYamlFile
-      ? yaml.parse(encodedContent)
-      : JSON.parse(encodedContent);
   } catch (err) {
     throw err;
   }
@@ -47,9 +105,13 @@ async function stringify(obj, options = {}) {
     const toYaml = options.format !== 'json' && (!options.hasOwnProperty('json') || options.json !== true);
 
     if (toYaml) {
+      // Set YAML options
+      const yamlOptions = {}
+      yamlOptions.lineWidth = (options.lineWidth && options.lineWidth === -1 ? Infinity : options.lineWidth) || Infinity;
+
       // Convert Object to YAML string
-      const lineWidth = (options.lineWidth && options.lineWidth === -1 ? Infinity : options.lineWidth) || Infinity;
-      output = yaml.safeStringify(obj, {lineWidth});
+      output = yaml.safeStringify(obj, yamlOptions);
+      output = addQuotesToRefInString(output);
 
       // Decode large number YAML values safely before writing output
       output = decodeLargeNumbers(output);
@@ -203,12 +265,97 @@ function decodeLargeNumbers(output, isJson = false) {
   }
 }
 
+/**
+ * Add quotes to $ref in string
+ * @param yamlString YAML string.
+ * @returns {*} YAML string with quotes.
+ */
+function addQuotesToRefInString(yamlString) {
+  return yamlString.replace(/(\$ref:\s*)([^"'\s]+)/g, '$1"$2"');
+}
+
+/**
+ * Analyze the OpenAPI document.
+ * @param {object} oaObj - The OpenAPI document as a JSON object.
+ * @returns {{operations: *[], methods: any[], paths: *[], flags: any[], operationIds: *[], flagValues: any[], responseContent: any[], tags: any[]}}
+ */
+function analyzeOpenApi(oaObj) {
+  const flags = new Set();
+  const tags = new Set();
+  const operationIds = [];
+  const paths = [];
+  const methods = new Set();
+  const operations = [];
+  const responseContent = new Set();
+  const flagValues = new Set();
+
+  if (oaObj && oaObj.paths) {
+    Object.keys(oaObj.paths).forEach((path) => {
+      paths.push(path);
+      const pathItem = oaObj.paths[path];
+
+      Object.keys(pathItem).forEach((method) => {
+        methods.add(method.toUpperCase());
+        const operation = pathItem[method];
+        operations.push(`${method.toUpperCase()}::${path}`);
+
+        if (operation?.tags && Array.isArray(operation.tags)) {
+          operation.tags.forEach((tag) => {
+            if (tag.startsWith('x-')) {
+              flags.add(tag);
+            } else {
+              tags.add(tag);
+            }
+          });
+        }
+
+        if (operation?.operationId) {
+          operationIds.push(operation.operationId);
+        }
+
+        if (operation?.responses) {
+          Object.values(operation.responses).forEach((response) => {
+            if (response?.content) {
+              Object.keys(response.content).forEach((contentType) => {
+                responseContent.add(contentType);
+              });
+            }
+          });
+        }
+
+        Object.keys(operation).forEach((key) => {
+          if (key.startsWith('x-')) {
+            flagValues.add(`${key}: ${operation[key]}`);
+          }
+        });
+      });
+    });
+  }
+
+  return {
+    methods: Array.from(methods),
+    tags: Array.from(tags),
+    operationIds,
+    flags: Array.from(flags),
+    flagValues: Array.from(flagValues),
+    paths,
+    operations,
+    responseContent: Array.from(responseContent),
+  };
+}
+
 module.exports = {
+  parseString,
   parseFile,
+  isJSON,
+  isYaml,
+  detectFormat,
   stringify,
   writeFile,
   encodeLargeNumbers,
   decodeLargeNumbers,
   getLocalFile,
   getRemoteFile,
+  analyzeOpenApi,
+  addQuotesToRefInString
 };